Revert "feat(markdown): save markdown files from UI"
