--- examples/ui/backend/utils/exceptions.py
@@ -1,6 +1,0 @@
-class FileNotFoundError(Exception):
-    pass
-
-
-class RestrictedAccessError(Exception):
-    pass
