+++ examples/ui/frontend/package.json
@@ -10,7 +10,6 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^5.1.1",
-    "@radix-ui/react-dialog": "^1.1.14",
     "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-slot": "^1.2.3",
     "class-variance-authority": "^0.7.1",
